Add tests coverage, better move generation, major bugfixes

+++ src/evaluation.cpp
@@ -113,7 +113,7 @@ int score_move(const game_t* game,
 {
   if (tt_move != 0 && move == tt_move) { return ORDER_TT_MOVE; }
 
-  assert(MOVE_PROMOTED(move) <
+  assert(static_cast<size_t>(MOVE_PROMOTED(move)) <
   sizeof(piece_values_abs) / sizeof(piece_values_abs[0]));
   
   // Promotion_t maps onto W_KNIGHT..W_QUEEN by construction.
